requirements.txt
requests
Pillow
pandas
openpyxl
xlsxwriter
tkcalendar
pymongo